+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.8 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- HeroBanner can't be a folder default view
+  [boulch]
 
 
 1.0.7 (2022-04-25)